Everything Recovered After HDD Failure
This is the code that I printed a while before the project got lost.
The code for the transmission cycle and setting of the power output
will need to be rewritten.

I also added the relevant manuals and photos, together with a short
README file that explains what the project is about.